tests/requirements.txt
pytest==3.0.1
pytest-ansible
pytest-cache
pytest-capturelog
pytest-cov
pytest-github
pytest-html
pytest-instafail
pytest-ordering
pytest-selenium
python-dateutil
git+ssh://git@github.com/ansible/towerkit.git
